requirements/base.txt
rules==2.0.1
psycopg2==2.8.3 --no-binary psycopg2  # https://github.com/psycopg/psycopg2
Werkzeug==0.14.1  # pyup: < 0.15 # https://github.com/pallets/werkzeug
pytz==2019.1  # https://github.com/stub42/pytz
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==4.1.2  # https://github.com/evansd/whitenoise

# Django
# ------------------------------------------------------------------------------
django==2.2.2  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django_import_export==1.2.0
django-extensions==2.1.9  # https://github.com/django-extensions/django-extensions

# Django REST Framework
djangorestframework==3.9.4  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
